Fix checksum result for assets already checksummed with another algorithm
The default checksum implementation remembers checksums it has already
performed so references to the same file do not need to be re-hashed
multiple times. However, this list was keyed on the assetPath alone.

With the introduction of configurable hash algorithms, it is possible
(in the unit tests, for example) for the same asset to be checksummed
using different algorithms at different times. Previously, the second
instance would just have the previous algorithm's result returned.

In particular, the introduction of checksum unit tests revealed this
issue. One test checks a default ('md5') checksum on `hours.jpg`
and another checks a specified `sha1` checksum on `hours.jpg`. Previously,
the `sha1` result was just returning the remembered `md5` checksum
since the `checksum` array was only keyed on the asset path.

This commit includes the hash algorithm in the key associated with the
remembered checksum. This ensures that requests to cachebust an asset
will remember previously cached checksums on a per-hashAlgorithm basis.